feat: add cli verbose option
