+++ packages/uns-cli/src/baseCommand.ts
@@ -58,6 +58,35 @@ export abstract class BaseCommand extends Command {
         }
     }
 
+    /**
+     * Enables this.log on every BaseCommand sub commands
+     */
+    public log(message = "", ...args: any[]): void {
+        // If help flag is set, we force logger. We can only test here.
+        if (this.verbose || this._helpOverride()) {
+            if (args && args.length > 0) {
+                super.log(message, args);
+            } else {
+                super.log(message);
+            }
+        }
+    }
+
+    /**
+     * Override of _helpOverride to take care of all help flags
+     */
+    public _helpOverride() {
+        for (const arg of this.argv) {
+            if (arg === "--help" || arg === "-h" || arg === "help") {
+                return true;
+            }
+            if (arg === "--") {
+                return false;
+            }
+        }
+        return false;
+    }
+
     protected getAvailableFormats(): Formater[] {
         return [OUTPUT_FORMAT.json];
     }
