+++ packages/uns-cli/src/commands/status.ts
@@ -1,9 +1,8 @@
 import { BaseCommand } from "../baseCommand";
-import { BaseCommandLogs } from "../baseCommandLogs";
 import { CommandOutput, Formater, OUTPUT_FORMAT } from "../formater";
 import { getNetworksListListForDescription } from "../utils";
 
-export class StatusCommand extends BaseCommandLogs {
+export class StatusCommand extends BaseCommand {
     public static description = "Display blockchain status";
 
     public static examples = [`$ uns status --network ${getNetworksListListForDescription()}`];
