+++ packages/uns-cli/src/readCommand.ts
@@ -6,12 +6,6 @@ export abstract class ReadCommand extends BaseCommand {
         /**
          * CONTEXT
          */
-        this.log("\nCONTEXT:");
-        this.logAttribute("network", this.api.network.name);
-        this.logAttribute("node", this.api.getCurrentNode());
-        this.logAttribute("readDateTime", chainmeta.timestamp.human);
-        this.logAttribute("height", chainmeta.height);
-
         return {
             chainmeta: {
                 network: this.api.network.name,
